requirements.txt
Flask==0.10.1
Flask-PyMongo==0.3.0
Flask-WTF==0.9.4
Jinja2==2.7.2
MarkupSafe==0.18
WTForms==1.0.5
Werkzeug==0.9.4
argparse==1.2.1
beautifulsoup4==4.3.2
coverage==3.7.1
distribute==0.6.34
fake-factory==0.3.2
flask-mongoengine==0.7.0
itsdangerous==0.23
lxml==3.3.1
mongoengine==0.8.7
nose==1.3.0
nose-pathmunge==0.1.2
pymongo==2.6.3
python-dateutil==2.2
six==1.5.2
wsgiref==0.1.2
